Delete Confirmation window added

+++ src/main/java/com/example/tourplanner/views/MainViewController.java
@@ -11,7 +11,6 @@ import javafx.scene.Node;
 import javafx.scene.control.Label;
 import javafx.scene.control.ListView;
 import javafx.scene.control.TextField;
-import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
 import javafx.scene.control.*;
 import javafx.scene.control.cell.PropertyValueFactory;
@@ -85,7 +84,9 @@ public class MainViewController implements Initializable {
      * Delete tour action
      */
     public void deleteTourAction(){
-        mainViewModel.deleteTour(tourListView.getSelectionModel().getSelectedItem());
+        if(AssertView.deleteConfirmation()){
+            mainViewModel.deleteTour(tourListView.getSelectionModel().getSelectedItem());
+        }
     }
 
     /**
